--- authentication-service/src/main/java/com/xml/model/AdditionalBill.java
@@ -1,45 +1,0 @@
-package com.xml.model;
-
-import javax.persistence.*;
-
-@Entity
-public class AdditionalBill {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "customer_id", nullable = false)
-    private User customer;
-
-    @Column
-    private float price;
-
-    public AdditionalBill() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public User getCustomer() {
-        return customer;
-    }
-
-    public void setCustomer(User customer) {
-        this.customer = customer;
-    }
-
-    public float getPrice() {
-        return price;
-    }
-
-    public void setPrice(float price) {
-        this.price = price;
-    }
-}
