+++ codebook-service/src/main/resources/application.properties
@@ -14,7 +14,7 @@ spring.datasource.initialization-mode=always
 # Hibernate ddl auto (create, create-drop, update): with "create-drop" the database
 # schema will be automatically created afresh for every start of application
 spring.jpa.properties.hibernate.globally_quoted_identifiers=true
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
 
 # Allows Hibernate to generate SQL optimized for a particular DBMS
