+++ message-service/src/main/resources/data.sql
@@ -1,0 +1,7 @@
+/* messages */
+INSERT INTO `xml`.`message` (`id`,`message`,`message_date`,`receiver_id`,`sender_id`) VALUES ('1','cao','2020-05-06 19:48:40.614962','2','1');
+INSERT INTO `xml`.`message` (`id`,`message`,`message_date`,`receiver_id`,`sender_id`) VALUES ('2','cao','2020-05-06 20:06:40.614962','1','2');
+INSERT INTO `xml`.`message` (`id`,`message`,`message_date`,`receiver_id`,`sender_id`) VALUES ('3','sta radis?','2020-05-06 20:08:40.614962','2','1');
+INSERT INTO `xml`.`message` (`id`,`message`,`message_date`,`receiver_id`,`sender_id`) VALUES ('4','nista','2020-05-06 20:10:40.614962','1','2');
+INSERT INTO `xml`.`message` (`id`,`message`,`message_date`,`receiver_id`,`sender_id`) VALUES ('5','ti?','2020-05-06 20:10:40.614962','1','2');
+INSERT INTO `xml`.`message` (`id`,`message`,`message_date`,`receiver_id`,`sender_id`) VALUES ('6','nista','2020-05-06 20:18:40.614962','2','1');
