+++ rent-request-service/src/main/java/com/xml/model/AdditionalBill.java
@@ -1,0 +1,44 @@
+package com.xml.model;
+
+import javax.persistence.*;
+
+@Entity
+public class AdditionalBill {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column( nullable = false)
+    private Long customerId;
+
+    @Column
+    private float price;
+
+    public AdditionalBill() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Long getCustomerId() {
+        return customerId;
+    }
+
+    public void setCustomerId(Long customerId) {
+        this.customerId = customerId;
+    }
+
+    public float getPrice() {
+        return price;
+    }
+
+    public void setPrice(float price) {
+        this.price = price;
+    }
+}
