+++ rent-request-service/src/main/java/com/xml/model/RentRequest.java
@@ -20,11 +20,10 @@ public class RentRequest {
     @Column(nullable = false)
     private LocalDateTime reservedTo;
 
-   // @ManyToMany
-   // @JoinTable(name = "rented_advertisements", joinColumns = @JoinColumn(name = "rent_request_id", referencedColumnName = "id"),
-   //        inverseJoinColumns = @JoinColumn(name = "advertisement_id", referencedColumnName = "id"))
-  //  @Column
-  //  private Set<Long> advertisementsForRent;
+    @ElementCollection
+    @Column(name = "advertisement_id")
+    @CollectionTable(name = "rented_advertisements", joinColumns = @JoinColumn(name = "rent_request_id"))
+    private Set<Long> advertisementsForRent;
 
     @Column(name = "status")
     @Enumerated(EnumType.STRING)
@@ -61,14 +60,6 @@ public class RentRequest {
         this.reservedTo = reservedTo;
     }
 
- /*   public Set<Long> getAdvertisementsForRent() {
-        return advertisementsForRent;
-    }
-
-    public void setAdvertisementsForRent(Set<Long> advertisementsForRent) {
-        this.advertisementsForRent = advertisementsForRent;
-    }
-*/
     public RentRequestStatus getRentRequestStatus() {
         return rentRequestStatus;
     }
@@ -91,5 +82,13 @@ public class RentRequest {
 
     public void setReports(Set<Report> reports) {
         this.reports = reports;
+    }
+
+    public Set<Long> getAdvertisementsForRent() {
+        return advertisementsForRent;
+    }
+
+    public void setAdvertisementsForRent(Set<Long> advertisementsForRent) {
+        this.advertisementsForRent = advertisementsForRent;
     }
 }
