+++ rent-request-service/src/main/java/com/xml/model/Report.java
@@ -9,9 +9,6 @@ public class Report {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column
-    private Long carId;
-
     @ManyToOne
     @JoinColumn(name = "rent_request_id", nullable = false)
     private RentRequest rentRequest;
@@ -25,8 +22,7 @@ public class Report {
     public Report() {
     }
 
-    public Report(Long carId, float km, String additionalInformation) {
-        this.carId = carId;
+    public Report(float km, String additionalInformation) {
         this.km = km;
         this.additionalInformation = additionalInformation;
     }
@@ -37,14 +33,6 @@ public class Report {
 
     public void setId(Long id) {
         this.id = id;
-    }
-
-    public Long getCarId() {
-        return carId;
-    }
-
-    public void setCarId(Long carId) {
-        this.carId = carId;
     }
 
     public float getKm() {
