+++ zuul-api-gateway-service/src/main/resources/bootstrap.properties
@@ -9,7 +9,7 @@ zuul.routes.message-service.sensitive-headers=Cookie,Set-Cookie
 
 server.port=8443
 server.ssl.enabled=true
-server.ssl.key-store= zuul-api-gateway-service/src/main/resources/bootsecurity.p12
+server.ssl.key-store= classpath:bootsecurity.p12
 server.ssl.key-store-password= nikolina
 server.ssl.keyStoreType= PKCS12
 server.ssl.keyAlias= bootsecurity
